requirements.txt
torch
torchvision
onnx
onnxruntime
numpy
Pillow
opencv-python
cerebrium
fastapi
uvicorn
requests
pytest